Fix issue when the exit confirmation is shown the selected answers get cleared
